Add HeaderStyles file to store styles for header component
